+++ packages/graphql-tool-utilities/src/polyfills.ts
@@ -1,3 +1,0 @@
-// apollo-codegen uses Object.entries and Array.flatmap
-import 'core-js/modules/es.object.entries';
-import 'core-js/modules/es.array.flat-map';