+++ packages/graphql-tool-utilities/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Breaking Change
+
+- Update minimum supported node version to 12.14.0. Add engines field to help enforce usage of this version. [#1906](https://github.com/Shopify/quilt/pull/1906)
+- Removed core-js dependency, as its polyfills are no longer needed in node 12. [#1906](https://github.com/Shopify/quilt/pull/1906)
 
 ## 1.4.2 - 2021-04-13